+++ project-management-app/src/components/Login/Register.tsx
@@ -1,5 +1,19 @@
-function registerForm() {
-	return <>form to register new user:</>;
+import SignupForm from "./RegisterForm";
+
+function openRegister() {
+	console.log("Have to register");
 }
 
-export default registerForm;
+function Register() {
+	return (
+		<div className="main">
+			<SignupForm />
+			{/*<h1>If you are not registered click 🗺️: </h1>
+			<button id="register" onClick={openRegister}>
+				Register
+			</button>*/}
+		</div>
+	);
+}
+
+export default Register;
